Changes to long_term_scheduler
Moved the main long_term_scheduler functions to a new long_term_scheduler.cpp file.
